refactor(engine): extract models::manifest
Every ModelFile literal, every const table and all six table-building
macros land in one file, textually in their original order and each macro
still above the const it expands: `.github/scripts/check-model-plan-sizes.ts`
expands `macro_rules!` over one concatenated source, and a table split
across files would silently drop entries from the size canary.

`manifest_tests` moves whole rather than by subject — it is what
`cargo test models::manifest::manifest_tests` runs for a pin bump, and
splitting it would quietly shrink that command's coverage. It keeps all 10
tests; the five write_verified/verify_sha256 cases reach back into the
parent by name.

The imports carry cfgs the shared block in models.rs never needed: nothing
outside `validate_tts_langs` uses `coded_bail`/`ErrorCode` here, and `Path`
is only reachable on system_kokoro.

Refs #950
